Various fixes and improvements (#269)
- Bump version for next release
- Fix default gamma for CYD 2432S032C
- Remember gamma curve setting from Display settings app
- Add UART to Core2 (still has no voltage on Grove port, though)
- LVGL performance improvements: pin to second core and set task priority to "critical"
- Fix build warnings, including deprecations
- Removed deprecated `Thread` constructor
- Fix WaveShare S3 display: Some displays would show a white screen at 12MHz, so I'm putting it back to the
official config values.

+++ Tactility/Source/lvgl/LvglSync.cpp
@@ -22,7 +22,7 @@ void syncSet(LvglLock lock, LvglUnlock unlock) {
     auto old_unlock = unlock_singleton;
 
     // Ensure the old lock is not engaged when changing locks
-    old_lock(portMAX_DELAY);
+    old_lock((uint32_t)portMAX_DELAY);
     lock_singleton = lock;
     unlock_singleton = unlock;
     old_unlock();
